re-intergrate fix for TeX error about inf. glue
The TeX error "Infinite glue shrinkage found in a paragraph" is due
to copying the nodelist with node.copy_list but that does not copy the 
glue spec nodes. So when changing a glue_spec node one needs to create
a new one. Otherwise the previously copied list gets affected.