HOTFIX for bang replacement in LUA_PATH
On windows/Lua, the lua search path gets mangled: a bang (!) will be 
replaced by the working directory of the binary. To prevent this,
a re-assignment to package.path is necessary. Though the correct solution
will to properly escape characters in a path or to remove require()
altogether. See http://lua-users.org/lists/lua-l/2012-08/msg00052.html
for more information.