Switch to jekyll + XML based documentation
The "text" documentation is markdown + textile and generated
with jekyll (https://github.com/mojombo/jekyll/wiki). The
XML based documentation is for the command reference and
generated with a simple XSLT script. Both are merged in
dist/manual

This is also the start of German _and_ English documentation.

To generate the documentation use the Rake task "doc". 
mkdoc and sdcontrol -d / -m are obsolete now and will be
removed