New XML parser "luxor"
With the new xml parser we can get rid of
libxml2 bindings as well as the lpeg based
xml parser. No need for a special binary anymore.

The new xml parser seems to be as fast as
the xml bindings.

The reason behind this change is that compilation
of luatex with libxml2 bindings is not trivial. I've
wasted a lot of time on this.