Decode html entities
We assume that decoding all html entities (such as
&uuml; does not harm the processing. 

Sometimes we have to parse HTML contents within XML.
Despite the fact that it is more or less always
non-well formed XML, we have some surprising
HTML entities in there. I's probably against all
XML rules to have 'em in there, we accept it.