Fix multi line centered text
On a paragraph with multi lines, only the last
line was centered. This was due to the fact
that we only looked for a parfillskip as the
second last node (no -10000 penalty for the 'hard'
linebreak).

We now remove everything until the last skip, which
can be a parfillskip or a self introduced
skip by an explicit line break (\n or <br/>)