Don't insert glue after a unary minus
We used to have glue after a hyphen so that 
compound words such as mailing-list can be
broken after the dash/hyphen. This also affected
numbers like -1234.45 because when calculating
the minimum width of the number (for table cells)
the minus sign was not taken into account.

The solution (heuristic) is to look if a 
dash/hyphen follows a glyph node. If true, then
insert glue, otherwise don't do anything special.