Bugfix for odd/even table head
When the table starts on a page where there is
not enough room for the table head, we move to the
next page. But for the logic of the table, we
are still on the first page. tablepart_absolute
reflects this and is increased only if there is
a real table page break.