Start asynchronous handling of text
Text was unable to break between pages and only simple column breaking
is possible. Multi page table are limited to guessing on what page a
table is on.

With the async module we can prepare contents (text, table) and ask
the subsystem to place parts of the contents depending on the current
available space.