Building nested Element structure fixed
When building nested Elements in a variable such as

<SetVariable variable="foo">
   <Copy-of select="$foo" />
   <Element name="sub">
     ...
   </Element>
</SetVariable>

and using this structure inside another Element like this

<Element name="whatever">
  <Copy-of select="$foo" />
</Element>

we have an intermediate table between the outer Element
and the inner Element. Which breaks when trying to save
the structure to disc.

When building the Element structure we now explicitly look
for the intermediate table and remove it.