Workaround for very high groups
The groups have a fixed page size. This leads to problems when the
contents of a group is very large (high). A page break will occur in the
middle of a group which is so bad.

The workaround is to make a 10 times larger page for the group.

Todo: fix the problem by doing what has to be done: dynamically
stretching the group height and width.