Fix for min_wd in tables
This supersedes / corrects the previous commit (15c88df) with
a proper fix for min_wd in tables. The problem was that calculating
the minimum width in Paragraph:min_width() took the wrong text format
into account. When the text format is ragged_*, the setting
tex.pdfadjustspacing is temporarily set to 0, giving a different
width than non-ragged_* text format.

QA/tests are for allowbreak in tables and min_wd calculation