spot colors revisited
This is a cleaner implementation. Every spot color that is used
gets added to all following page resources, so we don’t need
to keep track which colors are used on a particular page.

While this might be a small waste of space, we make our lives
much easier with this implementation.