Add internal communication between sp and luatex
In the future I can ask the server (sp) and the
client (LuaTeX) to do things for me, such as breaking
a line into a paragraph or executing complex regular
expression. This commit might be the most important commit
this year, let’s hope it also works on Linux and Windows,
as it has been tested only on a mac.