Use Go for HTML/doc generation
We use Go instead of XSLT for HTML generation.
It is slightly faster. It will be easier to maintain in the future.
Hopefully.

This commit also includes the dash docset generation with sqlite3
(now a dependency), which was formerly done with a python script.
(Dash is an offline-documentation browser for the Mac).

Fixes #50