Temporary workaround for broken XML / UTF8
See #65 for details. 

The file pubisher.status is an XML file that gets written by spinit.lua
function exit() and includes (error) messages. These messages can be in
a non-UTF8 encoding. The encoding should be fixed, as reading the 
XML file generates error messages. This is especially bad in the 
server mode, which relies on the publisher.status file for obtaining the 
status of a specific run.

This workaround removes the (error) messages.