Make temp directories always start with a non-0
This doesn't fix a problem in the publisher. But sometimes when IDs
are handled outside of the publisher, leading zeros can cause problems.
To avoid that, we set a prefix to the temp directory (ID).

This commit uses a digit, instead of a letter, as all other
characters are currently a digit.


See also c059f255eb5ea28360e3434f7c4e0b27b3e62414