Better wrapping text around images
This is far from complete, but enhances the situation in many ways.

The problem is that we can't see into the future (pages). A long 
paragraph stretching on multiple pages need to take the par shape into
account for all images placed on the next (and second next,...) pages.

paragraph:format() is called for each paragraph, but we don't really 
know where we are, as the final placement is done after everything 
has been formatted (vsplit). 

This needs an architectural change.