B: Incorrect line height calculation allocate/auto
This fixes a bug where the the line height was taken from the wrong
node in the paragraph. 

Now we look at the first glyph_node and get the size from there.

On empty paragraphs (<Paragraph><Value /></Paragraph>) there is no 
glyph_node and no need to format the paragraph anyway, so we exit early.

This fixes #101