Use go library for image caching
This allows the user to switch between the old 'lua' image caching
method which was fast, but never checked if the image is outdated
and the new method (default) which uses an HTTP-request to check
if the image needs to be refreshed.

This also enables https requests for images as the Lua is not
linked with openssl.