Proof of concept of Lua preprocessor
This allows the user to run a Lua script instead of an XProc filter.
The only helper function is currently 

runtime.run_saxon(stylesheet,src,out,paramstring)

and the only runtime variable is

runtime.projectdir

to get the directory with the publisher.cfg / layout.xml file.

This also includes Saxon-HE 9.8.0.4J from Saxonica