New Par mode and HTML processing
This replaces the old paragraph.lua with the following changes:

 * HTML code in the data is now handled by the new HTML parser
 * Text is collected and formatted as late as possible. No font / color 
   settings are applied after the text has been turned into a node list.
 * Much better kerning with <sub>, <sup>, <NoBreak>, line breaking 
   within <sub> and <sup>

Be aware that this breaks backward compatibility. Hopefully only by 
little means, but there are some cases to be beware of.

Say this data file for example:

<data>
  <h1>Some title</h1>
</data>

and this layout file:

<Layout xmlns="urn:speedata.de:2009/publisher/en"
	xmlns:sd="urn:speedata:2009/publisher/functions/en">

  <DefineFontfamily fontsize="10" leading="12" name="myfont">
    <Regular fontface="sans" />
  </DefineFontfamily>

  <Record element="data">
    <PlaceObject>
      <Textblock width="4">
        <Paragraph fontface="myfont">
          <Value select="."/>
				</Paragraph>
			</Textblock>
		</PlaceObject>
	</Record>
</Layout>

Will now result in an error, because the h1 tag in the data switches
to bold, which is not defined in the current font.