Rename fontface to fontfamily in text commands
This makes more sense since a font face in publisher terms is a
physical font file (or a reference to it) and a family consists
of up to four values for regular/bold/... 

These text commands always choose a family, not a single face.

font face will be removed in version 5!