Rewrite command Initial (not backwards compatible)
The Initial command was based on heuristics and had no reliable results.
The rewrite is backwards incompatible, but shifts the initial nicely
based on the height of the initial and the height of the surrounding
nodelist.

This also introduces finer control over padding top and padding left, 
both which can be negative.

This fixes #287