Better implementation of border collapse
This partly fixes #332 in a sense that horizontal border collapse
with fixed widths columns now have support for collapsing but the
widths should be adjusted.

This also fixes #260, probably also with problems of height changes.

The border-collapse is not correct at a page break. 

I will leave this as it is now.